More fixes (#147)
* fix: fix our match for CertificationAuthority class to use >= instead of > to solve NTAuthCertificates node

* chore: fix another potential case confusion

* fix: missing properties in generator

* fix: exclude group policy container objects from container ldap filter

* fix: use long property instead of int
chore: temporarily add enterprise dc group back in

* fix: keep unresolved principals in enrollment agent aces

+++ src/CommonLib/DirectoryObjects/DirectoryEntryWrapper.cs
@@ -89,7 +89,7 @@ public class DirectoryEntryWrapper : IDirectoryObject {
         return true;
     }
 
-    public bool TryGetIntProperty(string propertyName, out int value) {
+    public bool TryGetLongProperty(string propertyName, out long value) {
         value = 0;
         if (!CheckCache(propertyName)) return false;
 
@@ -97,7 +97,7 @@ public class DirectoryEntryWrapper : IDirectoryObject {
             return false;
         }
 
-        return int.TryParse(s, out value);
+        return long.TryParse(s, out value);
     }
 
     public bool TryGetCertificateArrayProperty(string propertyName, out X509Certificate2[] value) {
